Implement handle errors feature

+++ MQuince/MQuince.Integration.HospitalApp/Views/assets/js/PharmacyOffers.js
@@ -58,7 +58,9 @@ function deleteRow()
         
 };
 
-
+var negativePrice = false;
+var negativeQuant = false;
+var empty = false;
 function sendOffer() {
     var allMed = "";
     var allQuantity = "";
@@ -76,46 +78,83 @@ function sendOffer() {
         var quantity = document.getElementById("drugs").rows[i].getElementsByTagName('input')[1].value;
         var price = document.getElementById("drugs").rows[i].getElementsByTagName('input')[2].value;
 
+        if (pharmacyName == "" || pharmacyEmail == "" || medicationName == "" || quantity == "" || price == "") {
+            empty = true;
+            break;
+        }
+        else
+            empty = false;
+
         allMed += medicationName + ";";
         allQuantity += quantity + ";";
         allPrice += price + ";";
 
+       
+
+        if (price < 1) {
+
+            negativePrice = true;
+            break;
+        } else
+            negativePrice = false;
+
+        if (quantity < 1) {
+            negativeQuant = true;
+            break
+        }
+        else
+            negativeQuant = false;
+
+
+        
 
 
     }
     allMed = allMed.slice(0, -1);
     allQuantity = allQuantity.slice(0, -1);
     allPrice = allPrice.slice(0, -1);
+
+    if (empty) alert("Please fill all fields before submit!");
+    else if (negativePrice) {
+        alert("Please insert positive value for price!");
+    }
+    else if (negativeQuant)
+    {
+        alert("Please insert positive value for quantity")
+    }
+    else {
+        $.ajax({
+            url: "/api/PharmacyOffers",
+            type: "POST",
+            contentType: "application/json; charset=utf-8",
+            dataType: "json",
+            data: JSON.stringify({
+                IdTender: id.substring(1, id.length - 1),
+                PharmacyName: pharmacyName,
+                PharmacyEmail: phamracyEmail,
+                Medicationes: allMed,
+                Quantity: allQuantity,
+                Price: allPrice
+            }),
+
+            success: function (result) {
+                alert("Successefully send offers!");
+                localStorage.clear();
+                document.location.href = "http://localhost:49544/home";
+
+            },
+            error: function (data, ajaxOptions, thrownError) {
+                console.log('Error:', data);
+                alert("Cant be send!");
+                location.reload();
+            }
+        });
+
+        localStorage.clear();
+    }
     
 
-    $.ajax({
-        url: "/api/PharmacyOffers",
-        type: "POST",
-        contentType: "application/json; charset=utf-8",
-        dataType: "json",
-        data: JSON.stringify({
-            IdTender: id.substring(1, id.length - 1),
-            PharmacyName: pharmacyName,
-            PharmacyEmail: phamracyEmail,
-            Medicationes: allMed,
-            Quantity: allQuantity,
-            Price: allPrice
-        }),
-
-        success: function (result) {
-            alert("Successefully send offers!");
-            localStorage.clear();
-            document.location.href = "http://localhost:49544/home";
-           
-        },
-        error: function (data, ajaxOptions, thrownError) {
-            console.log('Error:', data);
-            alert("Cant be send!");
-            location.reload();
-        }
-    });
-
-    localStorage.clear();
+    
 }
